Only include custom OpenSSL ABI checks for OpenSSL 1.1
OpenSSL has a [Policy on API compatibility in minor releases][1] saying:

[1]: https://openssl-library.org/policies/technical/api-compat/

> Only API additions are allowed in minor releases.

Minor releases are defined as 3.X.0. A change to 4.X.X would change the
libssl.so.3 SONAME to libssl.so.4 so the (possibly) incompatible library would
not get loaded in the first place.

This warning message is causing confusion between unrelated issues in the Arch
Linux bug tracker.