Un-parenthesize checks wrt connection_edge_process_ordered_relay_cell()
Previously, one of these checks had the parentheses in the wrong
place, given an incorrect result.  The code is hard enough to read
that I refactored both instances to be more obviously right.

I've grepped for similar errors elsewhere, but didn't find them.

Fixed #41070.  Bug not in any released Tor.