KeepaliveIsolateSOCKSAuth: expire based on time of last _use_.
Formerly, we expired circuits affected by KeepaliveIsolateSOCKSAuth
based on the time when a stream was last attached to them, which would
sometimes result in circuits closing immediately after their last stream
was closed.  Now, we update the "dirty timestamp" on circuits whenever a
stream is removed, so that they'll expire based on the time when they
last had a circuit.

This implements the "easy version" of proposal 368.

Closes #41157.